requirements.txt
Django==2.0.2
configparser==3.5.0
gunicorn==19.7.1
django-webpack-loader==0.6.0
django-revproxy==0.9.13
requests==2.13.0

pytest==3.6.2
pytest-django==3.3.0
pytest-cov==2.5.1
pytest-mock==1.10.0
pytest-lazy-fixture==0.4.1
freezegun==0.3.10

celery==4.0.2
django-celery-beat==1.1.1
redis==2.10.6
kombu==4.1.0

psycopg2==2.7.3.1

django-oauth-toolkit==1.1.0
django-cors-middleware==1.3.1

django_serializer==0.0.6
django-admin-sortable2==0.6.19

phonenumberslite==8.9.2

django-pdfkit==0.3.1
django-colorfield==0.1.15
django-solo==1.1.3

django-mptt==0.9.1
git+https://github.com/RafaDias/django-smart-selects

Pillow==5.1.0
django-mailgun==0.9.1

django-jet==1.0.8